% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K Nearest Neighbors Cross Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Numeric input data to train and perform cross validation on.}

\item{cl}{String input that represents the column being predicted
(the true class).}

\item{k_nn}{Integer value that represents the number of neighbors.}

\item{k_cv}{Integer value that represents the number of yhat predictions.}
}
\value{
List value containing the average Cross Validation (misclass) error,
the predicted class values, and the training misclassification error.
}
\description{
This function takes input to perform a k-nearest neighbors cross validation
to suggest the best data model to use for prediction.
}
\examples{
my_knn_cv(data, "height", 1, 5)
my_knn_cv(data, "height", 3, 5)
my_knn_cv(data, "height", 5, 5)

}
\keyword{cross-validation,}
\keyword{prediction}
