% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Modeling Function}
\usage{
my_lm(formula, input)
}
\arguments{
\item{formula}{Formula input that describes the relationship
being explored in this model.}

\item{input}{Numeric input that contains the data used for this model.}
}
\value{
Table value with the estimated beta coefficient, std Error, test
statistic, and p-value.
}
\description{
This function will take data and a desired formula and output a
linear regression model.
}
\examples{
data_x <- c(1, 2, 3, 3, 2, 3, 2, 1, 1, 1, 2, 1, 2, 3, 3)
data_x2 <- c(10, 50, 30, 30, 20, 30, 20, 10, 10, 10, 20, 10, 20, 30, 30)
data_x3 <- c(13, 22, 43, 83, 42, 73, 22, 13, 11, 41, 62, 91, 22, 53, 53)
data_x4 <- c(14, 23, 132, 123, 572, 123, 242, 641, 231, 651, 232, 541, 232, 163, 833)
data_y <- c(13, 72, 33, 43, 72, 83, 32, 21, 15, 12, 62, 31, 27, 33, 36)

dataframe1 <- data.frame(data_x, data_x2, data_x3, data_x4, data_y)

my_lm(data_x3~data_x2, dataframe1)
my_lm(data_y~data_x3 + data_x2, dataframe1)
my_lm(data_y~data_x + data_x2 + data_x3, dataframe1)

}
\keyword{inference}
\keyword{regression,}
