% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R, R/my_penguins.R
\docType{data}
\name{my_penguins}
\alias{my_penguins}
\title{Measurements for life expectancy, population, and GDP from around the world.}
\format{
A tibble with 344 rows and 8 variables:
\describe{
\item{country}{a factor denoting country}
\item{continent}{a factor denoting continent (Asia, Europe, Africa, Americas, Oceania)}
\item{year}{a number denoting the year}
\item{lifeExp}{a number denoting life expectancy at birth (in years)}
\item{pop}{a number denoting the total population}
\item{gdpPercap}{a number denoting per-capita GDP}
}

A tibble with 344 rows and 8 variables:
\describe{
\item{species}{a factor denoting penguin species (Adélie, Chinstrap and Gentoo)}
\item{island}{a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
\item{bill_length_mm}{a number denoting bill length (millimeters)}
\item{bill_depth_mm}{a number denoting bill depth (millimeters)}
\item{flipper_length_mm}{an integer denoting flipper length (millimeters)}
\item{body_mass_g}{an integer denoting body mass (grams)}
\item{sex}{a factor denoting penguin sex (female, male)}
\item{year}{an integer denoting the study year (2007, 2008, or 2009)}
}
}
\source{
{Babies per Woman (total fertlity rate)} \url{https://www.gapminder.org/data/documentation/gd008/}

{Child Mortality Rate, under age five} \url{https://www.gapminder.org/data/documentation/gd005/}

{GDP per capita in constant PPP dollars} \url{https://www.gapminder.org/data/documentation/gd001/}

{Infant Mortality Rate, under age one} \url{https://www.gapminder.org/data/documentation/gd002/}

{Life Expectancy at Birth}\url{https://www.gapminder.org/data/documentation/gd004/}

{Population}\url{https://www.gapminder.org/data/documentation/gd003/}

{Average age at 1st marriage (girls)}\url{https://www.gapminder.org/data/documentation/gd009/}

{Gini}\url{http://gapm.io/doc_gini}

{HIV/AIDS}\url{https://www.gapminder.org/data/documentation/gd006/}

{Legal slavery}\url{https://www.gapminder.org/data/documentation/gd010/}

{Maternal mortality}\url{https://www.gapminder.org/data/documentation/gd010/}

{Income Mountains}\url{https://www.gapminder.org/data/documentation/income-mountains-dataset/}

{World Health Chart, data sources}\url{https://www.gapminder.org/data/documentation/gd000/}

{MORE DATA SOURCES ON GAPMINDER WEBSITE HERE:}\url{https://www.gapminder.org/data/}

{Adélie penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Adélie penguins (Pygoscelis adeliae) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}

{Gentoo penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Gentoo penguin (Pygoscelis papua) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/7fca67fb28d56ee2ffa3d9370ebda689}

{Chinstrap penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Chinstrap penguin (Pygoscelis antarcticus) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 6. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/c14dfcfada8ea13a17536e73eb6fbe9e}

{Originally published in: Gorman KB, Williams TD, Fraser WR (2014) Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis). PLoS ONE 9(3): e90081. doi:10.1371/journal.pone.0090081}
}
\usage{
my_penguins

my_penguins
}
\description{
Includes measurements and records of life expectancy, population, and
GDP for each country during multiple years.
This data and documentation come from the \code{gapminder} package,
available as \code{\link[gapminder]{gapminder}}.

Includes measurements for penguin species, island in Palmer Archipelago,
size (flipper length, body mass, bill dimensions), and sex.
This data and documentation come from the \code{palmerpenguins} package,
available as \code{\link[palmerpenguins]{penguins}}.
}
\keyword{datasets}
