% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation on Penguins Data}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input data representing the number of folds within data.}
}
\value{
Numeric value of the cross validation error.
}
\description{
Designed specifically for the Penguins Data in R, this function will
predict output body_mass_g using covariates bill_length_mm,
bill_depth_mm, and flipper_length_mm.
}
\examples{
my_rf_cv(5)
my_rf_cv(4)
my_rf_cv(3)

}
\keyword{cross-validation,}
\keyword{forest,}
\keyword{prediction}
