% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ttest.R
\name{my_ttest}
\alias{my_ttest}
\title{T Test function}
\usage{
my_ttest(x, alternative, mu)
}
\arguments{
\item{x}{Numeric input data to perform the t test on.}

\item{alternative}{String input to determine the direction of
the test: less, greater, or two.sided.}

\item{mu}{Numeric input of the population mean value.}
}
\value{
List value containing the test statistic, p value, degrees
of freedom, and the \code{alternative} type.
}
\description{
This function takes input to perform a t test on the desired data
with the desired parameters.
}
\examples{
my_ttest(c(1,2,3,4,5), "less", 0)
my_ttest(c(0,0,0,-13), "two.sided", -2)

}
\keyword{inference}
\keyword{test,}
