% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K Nearest Neighbors Cross Validation}
\usage{
my_knn_cv(train, true_cl, k_nn, k_cv)
}
\arguments{
\item{train}{Numeric input data to train and perform cross validation on.}

\item{true_cl}{String input that represents the column being predicted
(the true class).}

\item{k_nn}{Integer value that represents the number of neighbors.}

\item{k_cv}{Integer value that represents the number of yhat predictions.}
}
\value{
List value containing the average Cross Validation (misclass) error,
the predicted class values, and the training misclassification error.
}
\description{
This function takes input to perform a k-nearest neighbors cross validation
to suggest the best data model to use for prediction.
}
\examples{
data_x <- c(1, 2, 3, 3, 2, 3, 2, 1, 1, 1, 2, 1, 2, 3, 3)
data_x2 <- c(10, 50, 30, 30, 20, 30, 20, 10, 10, 10, 20, 10, 20, 30, 30)
data_x3 <- c(13, 22, 43, 83, 42, 73, 22, 13, 11, 41, 62, 91, 22, 53, 53)
data_x4 <- c(14, 23, 132, 123, 572, 123, 242, 641, 231, 651, 232, 541, 232, 163, 833)
data_y <- c(13, 72, 33, 43, 72, 83, 32, 21, 15, 12, 62, 31, 27, 33, 36)

dataframe1 <- data.frame(data_x, data_x2, data_x3, data_x4, data_y)

my_knn_cv(dataframe1, "data_y", 1, 5)
my_knn_cv(dataframe1, "data_x3", 3, 5)
my_knn_cv(dataframe1, "data_x2", 5, 5)

}
\keyword{cross-validation,}
\keyword{prediction}
