% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Modeling Function}
\usage{
my_lm(formula, input)
}
\arguments{
\item{formula}{Formula input that describes the relationship
being explored in this model.}

\item{input}{Numeric input that contains the data used for this model.}
}
\value{
Table value with the estimated beta coefficient, std Error, test
statistic, and p-value.
}
\description{
This function will take data and a desired formula and output a
linear regression model.
}
\examples{
my_lm(y~x, data)
my_lm(weight~height + gender, myData)
my_lm(temp~latitude + humidity + pollution, worldData)

}
\keyword{inference}
\keyword{regression,}
